move message related function to msg_simcom and msg_app
