+++ src/simcom/msg_proc_simcom.h
@@ -12,7 +12,5 @@
 #include "session.h"
 
 int handle_simcom_msg(const char* m, size_t msgLen, void* arg);
-int simcom_msg_send(void *msg, size_t len, SESSION *session);
-
 
 #endif /* SRC_MSG_PROC_SIMCOM_H_ */
